chore: Gossip state summary is typed
