+++ crates/api/src/gossip.rs
@@ -5,8 +5,10 @@ use crate::peer_store::DynPeerStore;
 use crate::transport::DynTransport;
 use crate::{
     builder, config, BoxFut, DynLocalAgentStore, DynOpStore, DynPeerMetaStore,
-    K2Result, SpaceId, StoredOp,
+    K2Result, SpaceId, StoredOp, Timestamp, Url,
 };
+use serde::{Deserialize, Serialize};
+use std::collections::HashMap;
 use std::sync::Arc;
 
 /// Request for a gossip state summary.
@@ -29,7 +31,7 @@ pub trait Gossip: 'static + Send + Sync + std::fmt::Debug {
     fn get_state_summary(
         &self,
         request: GossipStateSummaryRequest,
-    ) -> BoxFut<'_, K2Result<serde_json::Value>>;
+    ) -> BoxFut<'_, K2Result<GossipStateSummary>>;
 }
 
 /// Trait-object [Gossip].
@@ -61,3 +63,57 @@ pub trait GossipFactory: 'static + Send + Sync + std::fmt::Debug {
 
 /// Trait-object [GossipFactory].
 pub type DynGossipFactory = Arc<dyn GossipFactory>;
+
+/// DHT segment state.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct DhtSegmentState {
+    /// The top hash of the DHT ring segment.
+    pub disc_top_hash: bytes::Bytes,
+    /// The boundary timestamp of the DHT ring segment.
+    pub disc_boundary: Timestamp,
+    /// The top hashes of each DHT ring segment.
+    pub ring_top_hashes: Vec<bytes::Bytes>,
+}
+
+/// Peer metadata dump.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct PeerMeta {
+    /// The timestamp of the last gossip round.
+    pub last_gossip_timestamp: Option<Timestamp>,
+    /// The bookmark of the last op bookmark received.
+    pub new_ops_bookmark: Option<Timestamp>,
+    /// The number of behavior errors observed.
+    pub peer_behavior_errors: Option<u32>,
+    /// The number of local errors.
+    pub local_errors: Option<u32>,
+    /// The number of busy peer errors.
+    pub peer_busy: Option<u32>,
+    /// The number of terminated rounds.
+    ///
+    /// Note that termination is not necessarily an error.
+    pub peer_terminated: Option<u32>,
+    /// The number of completed rounds.
+    pub completed_rounds: Option<u32>,
+    /// The number of peer timeouts.
+    pub peer_timeouts: Option<u32>,
+}
+
+/// Gossip round state summary.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GossipRoundStateSummary {
+    /// The URL of the peer with which the round is initiated.
+    pub session_with_peer: Url,
+}
+
+/// Gossip state summary.
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct GossipStateSummary {
+    /// The current initiated round summary.
+    pub initiated_round: Option<GossipRoundStateSummary>,
+    /// The list of accepted round summaries.
+    pub accepted_rounds: Vec<GossipRoundStateSummary>,
+    /// DHT summary.
+    pub dht_summary: HashMap<String, DhtSegmentState>,
+    /// Peer metadata dump for each agent in this space.
+    pub peer_meta: HashMap<Url, PeerMeta>,
+}
