+++ crates/core/src/factories/core_gossip.rs
@@ -1,5 +1,5 @@
 use kitsune2_api::*;
-use serde_json::Value;
+use std::collections::HashMap;
 use std::sync::Arc;
 
 #[cfg(test)]
@@ -64,8 +64,15 @@ impl Gossip for CoreGossipStub {
     fn get_state_summary(
         &self,
         _request: GossipStateSummaryRequest,
-    ) -> BoxFut<'_, K2Result<Value>> {
-        Box::pin(async move { Ok(serde_json::json!({})) })
+    ) -> BoxFut<'_, K2Result<GossipStateSummary>> {
+        Box::pin(async move {
+            Ok(GossipStateSummary {
+                initiated_round: None,
+                accepted_rounds: Vec::with_capacity(0),
+                dht_summary: HashMap::with_capacity(0),
+                peer_meta: HashMap::with_capacity(0),
+            })
+        })
     }
 }
