+++ crates/gossip/src/lib.rs
@@ -11,9 +11,6 @@ pub use constant::*;
 mod gossip;
 pub use gossip::*;
 
-mod update;
-pub use summary::*;
-
 mod error;
 mod initiate;
 mod peer_meta_store;
@@ -23,3 +20,4 @@ mod state;
 mod storage_arc;
 mod summary;
 mod timeout;
+mod update;
